Refactored entire project and added a way to easly seperate games into diffrent classes

+++ include/springhawk/Time.h
@@ -8,13 +8,13 @@
 class Time {
 
 protected:
-    static void setDeltaTime(float dt);
+    static void setDeltaTime(double dt);
 
 public:
     friend class Engine;
-    static float getDeltaTime();
+    static double getDeltaTime();
 private:
-    static float deltaTime;
+    static double deltaTime;
 };
